requirements.txt
Flask~=2.0.2
pyAesCrypt~=6.0.0
psutil~=5.8.0
py-cpuinfo~=8.0.0
tqdm~=4.62.3